Moved some shared test utils to jest.setup.ts or '__mocks__/common.ts'
