+++ __mocks__/common.ts
@@ -1,0 +1,8 @@
+
+// Create a mock for scrollIntoView and focus
+export const mockScrollIntoView = jest.fn();
+
+// Mock window.scrollTo
+export const mockScrollTo = () => {
+  window.scrollTo = jest.fn();
+};
