+++ app/[locale]/account/profile/__tests__/page.spec.tsx
@@ -1,12 +1,17 @@
 import React from 'react';
-import {act, fireEvent, render, screen, within} from '@/utils/test-utils';
-import {axe, toHaveNoViolations} from 'jest-axe';
+import { act, fireEvent, render, screen, within } from '@/utils/test-utils';
+import { axe, toHaveNoViolations } from 'jest-axe';
 import ProfilePage from '../page';
 import {
   useLanguagesQuery,
   useMeQuery,
   useUpdateUserProfileMutation
 } from '@/generated/graphql';
+
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from '@/__mocks__/common';
 
 expect.extend(toHaveNoViolations);
 
@@ -26,11 +31,6 @@ jest.mock('@/generated/graphql', () => ({
   useLanguagesQuery: jest.fn(),
 }));
 
-jest.mock('@/context/ToastContext', () => ({
-  useToast: jest.fn(() => ({
-    add: jest.fn(),
-  })),
-}));
 
 // Mock UpdateEmailAddress component
 jest.mock('@/components/UpdateEmailAddress', () => ({
@@ -48,9 +48,6 @@ jest.mock('@/i18n/routing', () => ({
   usePathname: jest.fn(() => '/about'),
 }));
 
-jest.mock('@/components/PageHeader');
-
-
 // Mock useFormatter and useTranslations from next-intl
 jest.mock('next-intl', () => ({
   useFormatter: jest.fn(() => ({
@@ -59,9 +56,6 @@ jest.mock('next-intl', () => ({
   useTranslations: jest.fn(() => jest.fn((key) => key)), // Mock `useTranslations`,
   useLocale: jest.fn(() => 'en-US'), // Return a default locale
 }));
-
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
 
 const mockUserData = {
   me: {
@@ -91,7 +85,7 @@ describe('ProfilePage', () => {
   beforeEach(() => {
     setupMocks();
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
     const mockUpdateUserProfile = jest.fn();
     mockHook(useUpdateUserProfileMutation).mockReturnValue([mockUpdateUserProfile, { loading: false, error: undefined }]);
   });
