+++ app/[locale]/projects/[projectId]/__tests__/page.spec.tsx
@@ -1,60 +1,26 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { useParams } from 'next/navigation';
-import { useTranslations as OriginalUseTranslations } from 'next-intl';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import {
   useProjectQuery
 } from '@/generated/graphql';
 import ProjectOverviewPage from '../page';
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from "@/__mocks__/common";
 
 expect.extend(toHaveNoViolations);
-
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
 
 // Mock the useTemplateQuery hook
 jest.mock("@/generated/graphql", () => ({
   useProjectQuery: jest.fn(),
 }));
 
-// Mock next-intl
-jest.mock('next-intl', () => ({
-  useTranslations: () => (key: string) => key,
-}));
-
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
   useParams: jest.fn()
-}));
-
-type UseTranslationsType = ReturnType<typeof OriginalUseTranslations>;
-
-// Mock useTranslations from next-intl
-jest.mock('next-intl', () => ({
-  useTranslations: jest.fn(() => {
-    const mockUseTranslations: UseTranslationsType = ((key: string) => key) as UseTranslationsType;
-
-    /*eslint-disable @typescript-eslint/no-explicit-any */
-    mockUseTranslations.rich = (
-      key: string,
-      values?: Record<string, any>
-    ) => {
-      // Handle rich text formatting
-      if (values?.p) {
-        return values.p(key); // Simulate rendering the `p` tag function
-      }
-      return key;
-    };
-
-    return mockUseTranslations;
-  }),
-}));
-
-jest.mock('@/context/ToastContext', () => ({
-  useToast: jest.fn(() => ({
-    add: jest.fn(),
-  })),
 }));
 
 const mockProjectData = {
@@ -124,7 +90,7 @@ const mockProjectData = {
 describe('ProjectOverviewPage', () => {
   beforeEach(() => {
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
 
     const mockUseParams = useParams as jest.Mock;
     // Mock the return value of useParams
