+++ app/[locale]/projects/[projectId]/project/__tests__/page.spec.tsx
@@ -1,46 +1,18 @@
 import React from 'react';
 import { act, render, screen, fireEvent, within } from '@testing-library/react';
 import { useParams } from 'next/navigation';
-import { useTranslations as OriginalUseTranslations } from 'next-intl';
 import { useProjectQuery, useTopLevelResearchDomainsQuery, useUpdateProjectMutation } from '@/generated/graphql';
 import ProjectsProjectDetail from '../page';
 import { axe, toHaveNoViolations } from 'jest-axe';
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from "@/__mocks__/common";
+
 expect.extend(toHaveNoViolations);
 
 jest.mock('next/navigation', () => ({
   useParams: jest.fn(),
-}));
-
-jest.mock('@/context/ToastContext', () => ({
-  useToast: jest.fn(() => ({
-    add: jest.fn(),
-  })),
-}));
-
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
-
-type UseTranslationsType = ReturnType<typeof OriginalUseTranslations>;
-
-// Mock useTranslations from next-intl
-jest.mock('next-intl', () => ({
-  useTranslations: jest.fn(() => {
-    const mockUseTranslations: UseTranslationsType = ((key: string) => key) as UseTranslationsType;
-
-    /*eslint-disable @typescript-eslint/no-explicit-any */
-    mockUseTranslations.rich = (
-      key: string,
-      values?: Record<string, any>
-    ) => {
-      // Handle rich text formatting
-      if (values?.p) {
-        return values.p(key); // Simulate rendering the `p` tag function
-      }
-      return key;
-    };
-
-    return mockUseTranslations;
-  }),
 }));
 
 jest.mock('@/generated/graphql', () => ({
@@ -56,7 +28,7 @@ describe('ProjectsProjectDetail', () => {
 
   beforeEach(() => {
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
     mockUseParams.mockReturnValue({ projectId: '1' });
     mockUseProjectQuery.mockReturnValue({
       data: {
