+++ app/[locale]/template/[templateId]/__tests__/page.spec.tsx
@@ -8,6 +8,10 @@ import {
 
 import { useParams } from 'next/navigation';
 import TemplateEditPage from '../page';
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from "@/__mocks__/common";
 
 // Mock the useTemplateQuery hook
 jest.mock("@/generated/graphql", () => ({
@@ -21,21 +25,12 @@ jest.mock('next/navigation', () => ({
   useParams: jest.fn(),
 }))
 
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
-
 // Mock useFormatter and useTranslations from next-intl
 jest.mock('next-intl', () => ({
   useFormatter: jest.fn(() => ({
     dateTime: jest.fn(() => '01-01-2023'),
   })),
   useTranslations: jest.fn(() => jest.fn((key) => key)), // Mock `useTranslations`
-}));
-
-jest.mock('@/context/ToastContext', () => ({
-  useToast: jest.fn(() => ({
-    add: jest.fn(),
-  })),
 }));
 
 jest.mock('@/components/BackButton', () => {
@@ -73,7 +68,7 @@ const mockTemplateData = {
 describe("TemplateEditPage", () => {
   beforeEach(() => {
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
     const mockTemplateId = 123;
     const mockUseParams = useParams as jest.Mock;
