+++ app/[locale]/template/[templateId]/access/__tests__/page.spec.tsx
@@ -1,15 +1,18 @@
 import React from 'react';
-import {fireEvent, render, screen, waitFor} from '@/utils/test-utils';
-import {axe, toHaveNoViolations} from 'jest-axe';
+import { fireEvent, render, screen, waitFor } from '@/utils/test-utils';
+import { axe, toHaveNoViolations } from 'jest-axe';
 import TemplateAccessPage from '../page';
-import {useParams} from 'next/navigation';
-import {useTranslations as OriginalUseTranslations} from 'next-intl';
+import { useParams } from 'next/navigation';
 import {
   useAddTemplateCollaboratorMutation,
   useRemoveTemplateCollaboratorMutation,
   useTemplateCollaboratorsQuery
 } from '@/generated/graphql';
 import mockData from '../__mocks__/mockData.json';
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from "@/__mocks__/common";
 
 expect.extend(toHaveNoViolations);
 
@@ -36,42 +39,10 @@ jest.mock('@/components/PageHeader', () => ({
   default: () => <div data-testid="mock-page-header" />
 }));
 
-jest.mock('@/context/ToastContext', () => ({
-  useToast: jest.fn(() => ({
-    add: jest.fn(),
-  })),
-}));
-
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
-
-type UseTranslationsType = ReturnType<typeof OriginalUseTranslations>;
-
-// Mock useTranslations from next-intl
-jest.mock('next-intl', () => ({
-  useTranslations: jest.fn(() => {
-    const mockUseTranslations: UseTranslationsType = ((key: string) => key) as UseTranslationsType;
-
-    /*eslint-disable @typescript-eslint/no-explicit-any */
-    mockUseTranslations.rich = (
-      key: string,
-      values?: Record<string, any>
-    ) => {
-      // Handle rich text formatting
-      if (values?.p) {
-        return values.p(key); // Simulate rendering the `p` tag function
-      }
-      return key;
-    };
-
-    return mockUseTranslations;
-  }),
-}));
-
 describe('TemplateAccessPage', () => {
   beforeEach(() => {
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
     const mockTemplateId = 123;
     const mockUseParams = useParams as jest.Mock;
