+++ app/[locale]/template/[templateId]/history/__tests__/page.spec.tsx
@@ -6,7 +6,11 @@ import { MockedProvider } from '@apollo/client/testing';
 import { useParams, useRouter } from 'next/navigation';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import { handleApolloErrors } from "@/utils/gqlErrorHandler";
-import mockData from './mockedResponse.json'
+import mockData from './mockedResponse.json';
+import {
+    mockScrollIntoView,
+    mockScrollTo
+} from "@/__mocks__/common";
 
 expect.extend(toHaveNoViolations);
 
@@ -49,13 +53,11 @@ jest.mock('@/components/PageHeader', () => {
 });
 
 const mockUseRouter = useRouter as jest.Mock;
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
 
 describe('TemplateHistory', () => {
     beforeEach(() => {
         HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-        window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+        mockScrollTo();
 
         const mockTemplateId = 123;
         const mockUseParams = useParams as jest.Mock;
