+++ app/[locale]/template/[templateId]/section/new/__tests__/page.spec.tsx
@@ -3,9 +3,11 @@ import { act, fireEvent, render, screen, waitFor } from '@/utils/test-utils';
 import { useTemplateQuery, } from '@/generated/graphql';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import { useParams } from 'next/navigation';
-import { useTranslations as OriginalUseTranslations } from 'next-intl';
 import SectionTypeSelectPage from '../page';
-
+import {
+  mockScrollIntoView,
+  mockScrollTo
+} from "@/__mocks__/common";
 expect.extend(toHaveNoViolations);
 
 // Mock the useTemplateQuery hook
@@ -16,33 +18,6 @@ jest.mock("@/generated/graphql", () => ({
 jest.mock('next/navigation', () => ({
   useParams: jest.fn(),
 }))
-
-// Create a mock for scrollIntoView and focus
-const mockScrollIntoView = jest.fn();
-
-type UseTranslationsType = ReturnType<typeof OriginalUseTranslations>;
-
-
-// Mock useTranslations from next-intl
-jest.mock('next-intl', () => ({
-  useTranslations: jest.fn(() => {
-    const mockUseTranslations: UseTranslationsType = ((key: string) => key) as UseTranslationsType;
-
-    /*eslint-disable @typescript-eslint/no-explicit-any */
-    mockUseTranslations.rich = (
-      key: string,
-      values?: Record<string, any>
-    ) => {
-      // Handle rich text formatting
-      if (values?.p) {
-        return values.p(key); // Simulate rendering the `p` tag function
-      }
-      return key;
-    };
-
-    return mockUseTranslations;
-  }),
-}));
 
 jest.mock('@/components/BackButton', () => {
   return {
@@ -95,7 +70,7 @@ const mockTemplateData = {
 describe("SectionTypeSelectPage", () => {
   beforeEach(() => {
     HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-    window.scrollTo = jest.fn(); // Called by the wrapping PageHeader
+    mockScrollTo();
     const mockTemplateId = 123;
     const mockUseParams = useParams as jest.Mock;
